pyproject.toml
[build-system]
requires = ["setuptools>=65.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "newshive"
version = "0.2.0"
description = "Specification-driven content discovery and enrichment pipeline"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "newshive Contributors"}
]

dependencies = [
    "requests>=2.31.0",
    "beautifulsoup4>=4.12.0",
    "playwright>=1.40.0",
    "selenium>=4.15.0",
    "chromedriver-autoinstaller>=0.6.0",
    "O365>=2.0.0",
    "textual>=0.50.0",
    "rich>=13.0.0",
    "python-dotenv>=1.0.0",
    "langdetect>=1.0.9",
    "feedparser>=6.0.0",
    "python-dateutil>=2.8.0",
    "trafilatura>=1.6.0",
    "pyyaml>=6.0.0",
]

[project.optional-dependencies]
openai = [
    "openai>=1.0.0",
]
anthropic = [
    "anthropic>=0.7.0",
]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "mypy>=1.4.0",
    "ruff>=0.1.0",
    "openai>=1.0.0",
    "anthropic>=0.7.0",
]

[project.scripts]
newshive = "src.cli:main"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
disallow_incomplete_defs = false
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
strict_equality = true

[[tool.mypy.overrides]]
module = [
    "openai.*",
    "anthropic.*",
]
ignore_missing_imports = true

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = [
    "E",    # pycodestyle errors
    "W",    # pycodestyle warnings
    "F",    # Pyflakes
    "I",    # isort
    "N",    # pep8-naming
    "UP",   # pyupgrade
    "B",    # flake8-bugbear
    "A",    # flake8-builtins
    "C4",   # flake8-comprehensions
    "PIE",  # flake8-pie
    "RUF",  # Ruff-specific
]
ignore = [
    "E501",  # line too long (handled by line-length)
]

[tool.ruff.lint.isort]
known-first-party = ["src"]

[tool.ruff.lint.per-file-ignores]
"tests/**" = [
    "A002",  # Argument name shadows builtin
    "A003",  # Attribute name shadows builtin
]
